+++ src/dungeon/quest-monster-placer.cpp
@@ -4,6 +4,7 @@
 #include "monster-floor/place-monster-types.h"
 #include "monster/monster-info.h"
 #include "system/dungeon/quest-definition.h"
+#include "system/dungeon/quest-list.h"
 #include "system/floor/floor-info.h"
 #include "system/grid-type-definition.h"
 #include "system/monrace/monrace-definition.h"
