+++ src/dungeon/quest.cpp
@@ -12,6 +12,7 @@
 #include "player-status/player-energy.h"
 #include "player/player-status.h"
 #include "system/dungeon/quest-definition.h"
+#include "system/dungeon/quest-list.h"
 #include "system/floor/floor-info.h" // @todo 相互参照、将来的に削除する.
 #include "system/grid-type-definition.h"
 #include "system/item/item-entity.h"
