+++ src/floor/wild.cpp
@@ -29,6 +29,7 @@
 #include "system/dungeon/dungeon-definition.h"
 #include "system/dungeon/dungeon-list.h"
 #include "system/dungeon/quest-definition.h"
+#include "system/dungeon/quest-list.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/enums/terrain/building-type.h"
 #include "system/enums/terrain/terrain-tag.h"
