+++ src/info-reader/fixed-map-parser.cpp
@@ -17,18 +17,15 @@
 #include "player-info/class-info.h"
 #include "player-info/race-info.h"
 #include "player/player-realm.h"
-#include "system/angband-exceptions.h"
 #include "system/angband-system.h"
 #include "system/dungeon/quest-definition.h"
+#include "system/dungeon/quest-list.h"
 #include "system/floor/floor-info.h"
 #include "system/player-type-definition.h"
 #include "util/angband-files.h"
 #include "util/string-processor.h"
 #include "view/display-messages.h"
 #include "world/world.h"
-#include <algorithm>
-#include <fmt/format.h>
-#include <fstream>
 #include <string>
 
 static concptr variant = "ZANGBAND";
@@ -293,74 +290,4 @@ parse_error_type parse_fixed_map(PlayerType *player_ptr, std::string_view name, 
 
     angband_fclose(fp);
     return err;
-}
-
-static std::ifstream create_ifs_quest_list(std::string_view file_name)
-{
-    const auto path = path_build(ANGBAND_DIR_EDIT, file_name);
-    std::ifstream ifs(path);
-    if (ifs) {
-        return ifs;
-    }
-
-    constexpr auto fmt = _("ファイルが見つかりません ({})", "File is not found ({})");
-    THROW_EXCEPTION(std::runtime_error, fmt::format(fmt, file_name));
-}
-
-/*!
- * @brief ファイルからパースして作成したクエスト番号配列を返す
- * @param player_ptr プレイヤーへの参照ポインタ
- * @param file_name ファイル名
- * @return クエスト番号の配列
- */
-std::set<QuestId> parse_quest_info(std::string_view file_name)
-{
-    std::set<QuestId> key_list;
-    const auto push_set = [&key_list, &file_name](auto quest_id, auto line_num) {
-        if (quest_id == QuestId::NONE) {
-            return;
-        }
-
-        if (key_list.find(quest_id) != key_list.end()) {
-            constexpr auto fmt = _("重複したQuestID {} ({}の{}行目)", "Duplicated Quest Id {} ({} line {})");
-            THROW_EXCEPTION(std::runtime_error, fmt::format(fmt, enum2i(quest_id), file_name, line_num));
-        }
-
-        key_list.insert(quest_id);
-    };
-
-    auto ifs = create_ifs_quest_list(file_name);
-    std::string line;
-    auto line_num = 0;
-    while (std::getline(ifs, line)) {
-        line_num++;
-        if (line.empty() || line.starts_with('#')) {
-            continue;
-        }
-
-        line = utf8_to_local(line);
-        const auto tokens = str_split(line, ':', true);
-        switch (tokens[0][0]) {
-        case 'Q': {
-            const auto is_quest_none = _(tokens[1].starts_with('$'), !tokens[1].starts_with('$')) || (tokens[2] != "N");
-            const auto quest_number = is_quest_none ? QuestId::NONE : i2enum<QuestId>(std::stoi(tokens[1].substr(_(0, 1))));
-            push_set(quest_number, line_num);
-            break;
-        }
-        case '%': {
-            const auto key_list_recursive = parse_quest_info(tokens[1]);
-            key_list.insert(key_list_recursive.cbegin(), key_list_recursive.cend());
-            break;
-        }
-        default:
-            break;
-        }
-    }
-
-    if (ifs.bad() || (ifs.fail() && !ifs.eof())) {
-        constexpr auto fmt = _("ファイルの読み込みに失敗しました ({})", "Failed to read file ({})");
-        THROW_EXCEPTION(std::runtime_error, fmt::format(fmt, file_name));
-    }
-
-    return key_list;
 }
