+++ src/knowledge/knowledge-quests.cpp
@@ -15,6 +15,7 @@
 #include "system/artifact/artifact-definition.h"
 #include "system/dungeon/dungeon-record.h"
 #include "system/dungeon/quest-definition.h"
+#include "system/dungeon/quest-list.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/floor/floor-info.h"
 #include "system/item/item-entity.h"
