+++ src/load/load.cpp
@@ -43,6 +43,7 @@
 #include "system/angband-system.h"
 #include "system/angband-version.h"
 #include "system/dungeon/quest-definition.h"
+#include "system/dungeon/quest-list.h"
 #include "system/inner-game-data.h"
 #include "system/player-type-definition.h"
 #include "system/system-variables.h"
