+++ src/market/building-quest.cpp
@@ -4,6 +4,7 @@
 #include "info-reader/fixed-map-parser.h"
 #include "market/building-util.h"
 #include "system/dungeon/quest-definition.h"
+#include "system/dungeon/quest-list.h"
 #include "system/enums/dungeon/dungeon-id.h"
 #include "system/floor/floor-info.h"
 #include "system/floor/wilderness-grid.h"
