+++ src/system/dungeon/quest-definition.cpp
@@ -111,92 +111,3 @@ const MonraceDefinition &QuestType::get_bounty() const
 {
     return MonraceList::get_instance().get_monrace(this->r_idx);
 }
-
-QuestList QuestList::instance{};
-
-QuestList &QuestList::get_instance()
-{
-    return instance;
-}
-
-/*!
- * @brief クエストの初期化
- * @details ソフトウェア起動時ではパース関数が動作しないので、各種初期化シーケンス時に遅延初期化する
- */
-void QuestList::initialize()
-{
-    const auto quest_numbers = parse_quest_info(QUEST_DEFINITION_LIST);
-    QuestType quest{};
-    quest.status = QuestStatusType::UNTAKEN;
-    this->quests.emplace(QuestId::NONE, quest);
-    for (const auto q : quest_numbers) {
-        this->quests.emplace(q, quest);
-    }
-}
-
-void QuestList::reset_all()
-{
-    for (auto &[_, quest] : this->quests) {
-        quest.reset();
-    }
-}
-
-QuestType &QuestList::get_quest(QuestId id)
-{
-    return this->quests.at(id);
-}
-
-const QuestType &QuestList::get_quest(QuestId id) const
-{
-    return this->quests.at(id);
-}
-
-std::vector<QuestId> QuestList::get_sorted_quest_ids() const
-{
-    std::vector<QuestId> quest_ids;
-    std::transform(++this->quests.begin(), this->quests.end(), std::back_inserter(quest_ids), [](const auto &x) { return x.first; });
-    std::stable_sort(quest_ids.begin(), quest_ids.end(), [this](auto x, auto y) { return this->order_completed(x, y); });
-    return quest_ids;
-}
-
-void QuestList::set_defeated_monster(QuestId id, short numbers)
-{
-    this->quests.at(id).cur_num = numbers;
-}
-
-void QuestList::set_max_monster(QuestId id, short numbers)
-{
-    this->quests.at(id).max_num = numbers;
-}
-
-void QuestList::set_type(QuestId id, QuestKindType type)
-{
-    this->quests.at(id).type = type;
-}
-
-void QuestList::set_monrace_id(QuestId id, MonraceId monrace_id)
-{
-    this->quests.at(id).r_idx = monrace_id;
-}
-
-void QuestList::set_flags(QuestId id, uint32_t flags)
-{
-    this->quests.at(id).flags = flags;
-}
-
-bool QuestList::is_quest_equals(QuestId id, QuestKindType type) const
-{
-    return this->quests.at(id).type == type;
-}
-
-bool QuestList::is_bounty_valid(QuestId id) const
-{
-    return this->quests.at(id).get_bounty().is_valid();
-}
-
-bool QuestList::order_completed(QuestId id1, QuestId id2) const
-{
-    const auto &quest1 = this->get_quest(id1);
-    const auto &quest2 = this->get_quest(id2);
-    return (quest1.comptime != quest2.comptime) ? (quest1.comptime < quest2.comptime) : (quest1.level < quest2.level);
-}
